refactor(blorktools): glb binary buffer util
