+++ packages/blorktools/src/asset_debugger/modals/html-editor-modal/html-editor-modal.js
@@ -28,8 +28,6 @@ import {
     onGlbBufferUpdate
 } from './model-integration';
 import { 
-    getBinaryBufferForMesh,
-    associateBinaryBufferWithMesh,
     MESH_BINARY_EXTENSION, 
     MESH_INDEX_PROPERTY, 
     BINARY_DATA_PROPERTY
